requirements.txt
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
python-dotenv==0.20.0
python-telegram-bot==20.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
tinydb==4.8.2
ton-client-py==1.44.3.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
